Backend for Multiple Entry Points (#1411)
* Backend for Multiple Entry Points

Introduces the basic backend on the compiler for zero or more entry
points.  Entry points have been extended to lists for several functions,
with loopFunctions have been extended to take in entry points and
indices as appropriate, to allow for multiple entry points once the
frontend is expanded.  Several functions are currently being assumed to
have a single entry point for simplicity and provide a work in progress
commit.

* Progress on debugging fixes

* Tests passing

* Refactored emitEntryPoints

* Updated lists to be by constant reference

* Fixes to formatting

* Refactoring updates for the compiler

* Fix for compilation errors

* Reformatting

* More reformatting

* Moved struct around to help with compilation

Co-authored-by: Tim Foley <tfoleyNV@users.noreply.github.com>