Public Keyword for Functions (#1432)
This PR introduces support for the public modifier for functions.  This
keyword allows labelled functions to be written to the compiled without
having a link to an entry point.  The goal of this change is to help
support heterogeneous design of Slang by permitting C++ code to interact
with CPU slang functions.

Internally, this PR adds the public decoration to the IR and defines a
lowering from the public modifier in the AST to this decoration.
Additionally, the Keep Alive decoration is added to any public modifier
being lowered, which prevents DCE from eliminating functions labelled
with the public keyword.

Co-authored-by: Tim Foley <tfoleyNV@users.noreply.github.com>